+++ zork++/benches/benchmarks.rs
@@ -3,29 +3,23 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use zork::{
     cache::{self, ZorkCache},
+    cli::input::CliArgs,
     compiler::build_project,
     config_file::ZorkConfigFile,
     utils::{self, reader::build_model},
 };
 
-/// To succesfully run the benchmarks, change the filename of the
-/// [deactivated]zork_clang.toml under the github-example folder
 pub fn build_project_benchmark(c: &mut Criterion) {
     let config: ZorkConfigFile = toml::from_str(utils::constants::CONFIG_FILE_MOCK).unwrap();
     let program_data = build_model(&config);
+    let cache = ZorkCache::default();
 
     c.bench_function("Build project", |b| {
-        b.iter(|| {
-            build_project(
-                black_box(&program_data),
-                black_box(&ZorkCache::default()),
-                false,
-            )
-        })
+        b.iter(|| build_project(black_box(&program_data), black_box(&cache), false))
     });
 
     c.bench_function("Cache loading time", |b| {
-        b.iter(|| cache::load(black_box(&program_data)))
+        b.iter(|| cache::load(black_box(&program_data), &CliArgs::default()))
     });
 }