+++ zork++/src/lib/utils/reader.rs
@@ -61,7 +61,6 @@ pub fn find_config_files(base_path: &Path) -> Result<Vec<ConfigFile>> {
             && e.file_name().to_str().unwrap().starts_with("zork")
             && e.file_name().to_str().unwrap().ends_with(".toml")
         {
-            log::debug!("Config file found!: {:?}", e.file_name());
             files.push(ConfigFile {
                 dir_entry: e.clone(),
                 path: e.path().to_path_buf(),
@@ -242,8 +241,13 @@ fn assemble_module_implementation_model<'a>(
 ) -> ModuleImplementationModel<'a> {
     let mut dependencies = config.dependencies.clone().unwrap_or_default();
     if dependencies.is_empty() {
-        let implicit_dependency = config.file.split('.').collect::<Vec<_>>()[0];
-        dependencies.push(implicit_dependency);
+        let last_dot_index = config.file.rfind('.');
+        if let Some(idx) = last_dot_index {
+            let implicit_dependency = config.file.split_at(idx);
+            dependencies.push(implicit_dependency.0)
+        } else {
+            dependencies.push(config.file);
+        }
     }
 
     ModuleImplementationModel {